+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.2] - 2024-12-12
+
+### Fixed
+- `HttpConfig.appendUserAgent` now works as expected in the SDK rather than being overridden
+
+### Removed
+- `Application.Builder.removeCapabilities`, since it was added in Java SDK 8.14.0
+
 ## [1.1.1] - 2024-12-03
 - Bumped Java SDK version to 8.15.0